requirements.txt
numpy
py-tlsh
mysql-connector-python
pyyaml
matplotlib
seaborn
latex
ssdeep
sqlalchemy
psutil
brokenaxes